Refactor different parts into modules
